Implement class to generate projects.json
The fastest way to find repos within an organization that have a `.travis.yml` is to use GitHub's Search API. However, there are a couple of gotchas:

1. Some search queries will not return the same results consistently due to a known timeout issue. For example, this won't work:
Octokit.search_code 'travis in:path path:/ user:18F'

Instead, searching for the text "language" within a file called `travis.yml` seems to work consistently. Note that this returns duplicates for some reason, so we need to call `uniq` on the final array.

2. GitHub can't search forked repos unless the forked repo has more stars than the original repo. Therefore, we must process forked repos separately and use the Contents search, which is slower unfortunately.